pyproject.toml
[project]
name = "scalping-futures"
version = "0.1.0"
description = "scalping strategy for tinkoff.invest"
authors = [
    {name = "alexey.vasin",email = "alexeyvasin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0.0"

dependencies = [
    "tinkoff-investments (>=0.2.0b109,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
